+++ src/input/strParse.js
@@ -172,6 +172,34 @@ const strFmt = [
     }
   },
   {
+    // 'q2 2002'
+    reg: /^(q[0-9])( of)?( [0-9]{4})?/i,
+    parse: (s, arr) => {
+      let quarter = arr[1] || ''
+      s = s.quarter(quarter)
+      let year = arr[3] || ''
+      if (year) {
+        year = year.trim()
+        s = s.year(year)
+      }
+      return s
+    }
+  },
+  {
+    // 'summer 2002'
+    reg: /^(spring|summer|winter|fall|autumn)( of)?( [0-9]{4})?/i,
+    parse: (s, arr) => {
+      let season = arr[1] || ''
+      s = s.season(season)
+      let year = arr[3] || ''
+      if (year) {
+        year = year.trim()
+        s = s.year(year)
+      }
+      return s
+    }
+  },
+  {
     // '200bc'
     reg: /^[0-9,]+ ?b\.?c\.?$/i,
     parse: (s, arr) => {