+++ src/timezone/summerTime.js
@@ -8,12 +8,13 @@ const toUtc = (dstChange, offset, year) => {
 }
 
 // compare epoch with dst change events (in utc)
-const shouldChange = (epoch, start, end, summerOffset, winterOffset) => {
+const inSummerTime = (epoch, start, end, summerOffset, winterOffset) => {
   const year = new Date(epoch).getUTCFullYear()
   const startUtc = toUtc(start, winterOffset, year)
   const endUtc = toUtc(end, summerOffset, year)
+  // console.log(epoch, endUtc)
   // simple number comparison now
   return epoch >= startUtc && epoch < endUtc
 }
 
-module.exports = shouldChange
+module.exports = inSummerTime