+++ test/immutable.test.js
@@ -18,49 +18,49 @@ test('clone still works', (t) => {
 test('.add does not mutate', (t) => {
   t.plan(1)
   let tmrw = day0.add(1, 'day').format('nice')
-  t.ok(today === day0.format('nice'), 'not mutated')
+  t.ok(today === day0.format('nice'), '.add not mutated')
 })
 
 test('.subtract does not mutate', (t) => {
   t.plan(1)
   let tmrw = day0.subtract(1, 'day').format('nice')
-  t.ok(today === day0.format('nice'), 'not mutated')
+  t.ok(today === day0.format('nice'), '.subtract not mutated')
 })
 
 test('.hour does not mutate', (t) => {
   t.plan(1)
   let tmrw = day0.hour(1).format('nice')
-  t.ok(today === day0.format('nice'), 'not mutated')
+  t.ok(today === day0.format('nice'), '.hour not mutated')
 })
 
 test('.date does not mutate', (t) => {
   t.plan(1)
   let tmrw = day0.date(1).month(1).year(2018).format('nice')
-  t.ok(today === day0.format('nice'), 'not mutated')
+  t.ok(today === day0.format('nice'), '.date not mutated')
 })
 
 test('.day does not mutate', (t) => {
   t.plan(1)
   let tmrw = day0.day(22).format('nice')
-  t.ok(today === day0.format('nice'), 'not mutated')
+  t.ok(today === day0.format('nice'), '.day not mutated')
 })
 
 test('.month does not mutate', (t) => {
   t.plan(1)
   let tmrw = day0.month(7).format('nice')
-  t.ok(today === day0.format('nice'), 'not mutated')
+  t.ok(today === day0.format('nice'), '.month not mutated')
 })
 
 test('.quarter does not mutate', (t) => {
   t.plan(1)
   let tmrw = day0.quarter(4).format('nice')
-  t.ok(today === day0.format('nice'), 'not mutated')
+  t.ok(today === day0.format('nice'), '.quarter not mutated')
 })
 
 test('.goto does not mutate', (t) => {
   t.plan(1)
   let tmrw = day0.goto('Australia/Brisbane').format('nice')
-  t.ok(today === day0.format('nice'), 'not mutated')
+  t.ok(today === day0.format('nice'), '.goto not mutated')
 })
 test('time setting works', (t) => {
   t.equal(spacetime.now().time('6:00pm').time(), '6:00pm', 'input=output')