+++ test/season.test.js
@@ -31,3 +31,33 @@ test('season-by-hemisphere', (t) => {
   })
   t.end()
 })
+
+test('set season - north', (t) => {
+  let s = spacetime('winter', 'Canada/Eastern')
+  t.equal(s.monthName(), 'december', 'winter .month()')
+  t.equal(s.date(), 1, 'q1 .date()')
+
+  s = spacetime('spring', 'Canada/Eastern')
+  t.equal(s.monthName(), 'march', 'spring .month()')
+  t.equal(s.date(), 1, 'spring .date()')
+
+  s = spacetime('summer', 'Canada/Eastern')
+  t.equal(s.monthName(), 'june', 'summer .month()')
+  t.equal(s.date(), 1, 'summer .date()')
+
+  s = spacetime('fall', 'Canada/Eastern')
+  t.equal(s.monthName(), 'september', 'fall .month()')
+  t.equal(s.date(), 1, 'fall .date()')
+
+  s = spacetime('fall 2001', 'Canada/Eastern')
+  t.equal(s.monthName(), 'september', 'fall year .month()')
+  t.equal(s.date(), 1, 'fall year .date()')
+  t.equal(s.year(), 2001, 'fall .year()')
+
+  s = spacetime('fall of 1960', 'Canada/Eastern')
+  t.equal(s.monthName(), 'september', 'fall of year .month()')
+  t.equal(s.date(), 1, 'fall of year .date()')
+  t.equal(s.year(), 1960, 'fall of .year()')
+
+  t.end()
+})